ExodusDomain class added
